[LB-458] feat : 트레이싱 설정
