+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'com.lgcms'
-version = '0.5'
+version = '0.6'
 
 java {
 	toolchain {
@@ -39,8 +39,14 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.kafka:spring-kafka-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
+
+	implementation 'io.micrometer:micrometer-tracing-bridge-otel' // 트레이싱
+	implementation 'io.opentelemetry:opentelemetry-exporter-otlp' // otel exporter
+	implementation 'net.ttddyy.observation:datasource-micrometer-spring-boot:1.1.2' // JDBC 트레이싱
+	implementation 'io.github.openfeign:feign-micrometer' // feign 트레이싱
 }
 
 dependencyManagement {
