+++ src/main/resources/application-local.yml
@@ -5,44 +5,48 @@ spring:
   application:
     name: backend-lesson
   datasource:
-    url: jdbc:mysql://localhost:38121/lesson?serverTimezone=UTC&useSSL=true&autoReconnect=true&useUnicode=true&characterEncoding=utf-8
+    url: jdbc:postgresql://localhost:38121/lesson?serverTimezone=UTC&useSSL=true&autoReconnect=true&useUnicode=true&characterEncoding=utf-8
     username: lgcms
     password: 1234
-    driver-class-name: com.mysql.cj.jdbc.Driver
+    driver-class-name: org.postgresql.Driver
     hikari:
       connection-test-query: SELECT 1
       validation-timeout: 5000
 
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: none
     generate-ddl: true
     show-sql: true
     open-in-view: false
 
   kafka:
     bootstrap-servers: localhost:38123
-
+    template:
+      observation-enabled: true
     listener:
-      ack-mode: manual_immediate
+      observation-enabled: true
+      ack-mode: manual
 
     consumer:
       group-id: ${spring.application.name}
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      key-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
       enable-auto-commit: false
-      auto-offset-reset: latest
+      auto-offset-reset: earliest
       max-poll-records: 10
       properties:
+        spring.deserializer.key.delegate.class: org.apache.kafka.common.serialization.StringDeserializer
+        spring.deserializer.value.delegate.class: org.springframework.kafka.support.serializer.JsonDeserializer
         spring.json.trusted.packages: "*"
         spring.json.use.type.headers: false
+        spring.json.value.default.type: com.lgcms.lesson.common.kafka.dto.KafkaEvent
 
     producer:
       key-serializer: org.apache.kafka.common.serialization.StringSerializer
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
       properties:
         spring.json.add.type.headers: false
-
   cloud:
     openfeign:
       client:
